feat : 네이밍 변경 및 router 등록
