+++ src/views/wallet/create/LocalCardCreateDetailPage.vue
@@ -1,159 +1,107 @@
 <script setup lang="ts">
-import { ref, computed, watch, onMounted } from 'vue'
-import axios from 'axios'
 import Layout from '@/components/layout/Layout.vue'
+import DanjiChip from '@/components/common/chip/DanjiChip.vue'
 import DanjiButton from '@/components/common/button/DanjiButton.vue'
-import DanjiInput from '@/components/common/form/DanjiInput.vue'
-import DanjiSelect from '@/components/common/form/DanjiSelect.vue'
+import LocalFilterModal from '@/components/common/modal/LocalFilterModal.vue'
+import { onMounted, ref } from 'vue'
 import { useRoute } from 'vue-router'
-
-// 임시 타입
-interface Region {
-  id: number
-  name: string
-}
-
-interface City {
-  id: number
-  name: string
-}
-
-interface Card {
-  id: number
-  name: string
-}
-
-interface LocalData {
-  data: {
-    regions: Region[]
-    cities: Record<number, City[]>
-    cards: Record<number, Card[]>
-  }
-}
-
-const selectedRegion = ref('')
-const selectedCard = ref('')
-const password = ref('')
 
 const route = useRoute()
 
-const regions = ref<Region[]>([])
-const cities = ref<City[]>([])
-const cards = ref<Card[]>([])
+const isModalVisible = ref<boolean>(false)
+const selectedRegion = ref<string>('')
+const selectedCity = ref<string>('')
 
-const regionOptions = computed(() => regions.value.map((region) => region.name))
-const cardOptions = computed(() => cards.value.map((card) => card.name))
+const handleClickModal = (): void => {
+  isModalVisible.value = !isModalVisible.value
+}
+
+const handleModalConfirm = (region: string, city: string): void => {
+  selectedRegion.value = region
+  selectedCity.value = city
+  isModalVisible.value = false
+}
 
 // 라우터에서 전달받은 데이터 처리
 const initializeFromRoute = () => {
   const routeRegion = route.params.region
+  const routeCity = route.params.city
 
   if (routeRegion && typeof routeRegion === 'string') {
     selectedRegion.value = routeRegion
   }
-}
 
-// Mock API
-const fetchLocalData = async () => {
-  try {
-    const response = await axios.get<LocalData>('/api/local/list')
-    const data = response.data.data
-    regions.value = data.regions
-  } catch (error) {
-    console.error('데이터를 불러오는데 실패했습니다:', error)
+  if (routeCity && typeof routeCity === 'string') {
+    selectedCity.value = routeCity
   }
-}
-
-// 지역 변경 후 화폐 카테고리 업데이트
-watch(selectedRegion, async (newRegion) => {
-  if (!newRegion) return
-
-  try {
-    const response = await axios.get<LocalData>('/api/local/list')
-    const data = response.data.data
-
-    const region = data.regions.find((r) => r.name === newRegion)
-    if (region) {
-      cities.value = data.cities[region.id] || []
-      cards.value = data.cards[region.id] || []
-
-      // 카드 선택 초기화
-      selectedCard.value = ''
-
-      // 기본값으로 첫 번째 카드 설정
-      if (cards.value.length > 0) {
-        selectedCard.value = cards.value[0].name
-      }
-    }
-  } catch (error) {
-    console.error('지역 데이터를 불러오는데 실패했습니다:', error)
-  }
-})
-
-const handleSubmit = () => {
-  alert('카드 발급이 신청되었습니다.')
 }
 
 onMounted(async () => {
   initializeFromRoute()
-  await fetchLocalData()
+  console.log(selectedCity, selectedRegion)
 })
 </script>
 
 <template>
-  <Layout
-    header-type="basic"
-    header-title="지역화폐 카드 발급"
-    :show-left-icon="true"
-    :is-bottom-nav="false"
-  >
+  <layout header-type="basic" header-title="지역화폐 카드 발급" showLeftIcon :is-bottom-nav="false">
     <template #content>
-      <div class="h-full bg-Background overflow-hidden">
-        <div
-          class="px-[2.4rem] py-[3rem] mx-[1.6rem] space-y-[2.8rem] mt-[2rem] rounded-[1.6rem] bg-White-1 shadow-inner"
+      <!-- 상단 구분선 -->
+      <div class="w-full h-[0.1rem] bg-Gray-0" />
+
+      <!-- 지역 선택 칩 -->
+      <danji-chip
+        class="ms-[2.7rem] mt-[1.2rem]"
+        :is-active="true"
+        :is-icon="true"
+        @click="handleClickModal"
+      >
+        {{ selectedCity }}시
+      </danji-chip>
+
+      <!-- 지역 필터 모달 -->
+      <local-filter-modal
+        :is-visible="isModalVisible"
+        v-model:initial-region="selectedRegion"
+        v-model:initial-city="selectedCity"
+        @close="handleClickModal"
+        @confirm="handleModalConfirm"
+      />
+
+      <!-- 카드 이미지 임시 div -->
+      <div
+        class="mx-auto mt-[1.8rem] h-[19.8rem] w-[13.3rem] rounded-[0.8rem] border border-Gray-2 border-solid"
+      />
+
+      <!-- 카드 이름 -->
+      <div class="text-center mt-[0.4rem]">
+        <div class="text-Brown-4 Body01">부산 동백전</div>
+      </div>
+
+      <!-- 카드 혜택 정보 -->
+      <div class="flex flex-col mt-[3rem] mx-[2.7rem]">
+        <span class="Head03 text-Black-1"
+          >지역경제 활성화로 <br />
+          100(백)가지 행복과 즐거움을 주는 “동백전”</span
         >
-          <!-- 지역 선택 -->
-          <div>
-            <danji-select v-model="selectedRegion" label="지역 선택:" :options="regionOptions" />
-          </div>
 
-          <!-- 화폐 선택 -->
-          <div>
-            <danji-select v-model="selectedCard" label="화폐 선택:" :options="cardOptions" />
+        <span class="mt-[2rem] Head01 text-Black-1">Benefits</span>
+        <span class="mt-[0.3rem] Head04 text-Brown-2">동백전 인센티브 10%</span>
 
-            <!-- 혜택 정보 -->
-            <div class="ms-[0.6rem] mt-[0.2rem] space-y-[0.4rem]">
-              <p class="text-Brown-2 Head04">강릉페이 인센티브 10%</p>
-              <p>
-                <span class="text-Gray-5 Body02">최대 충전 가능 금액</span>
-                <span class="ms-[0.6rem] text-Yellow-1 Head04">600,000원</span>
-              </p>
-            </div>
-          </div>
-
-          <!-- 비밀번호 입력 -->
-          <div>
-            <label class="block mb-[0.4rem] text-Black-1 Head02"> 비밀번호 </label>
-            <danji-input
-              v-model="password"
-              type="password"
-              placeholder="지갑 비밀번호"
-              text-align="left"
-            />
-          </div>
+        <div>
+          <span class="Body02 text-Gray-5">최대 충전 가능 금액</span>
+          <span class="ms-[0.5rem] Head04 text-Yellow-1">600,000원</span>
         </div>
 
-        <!-- 발급하기 버튼 -->
-        <div class="mt-[2.6rem] flex justify-center">
-          <danji-button
-            variant="large"
-            :disabled="!selectedRegion || !selectedCard || !password"
-            @click="handleSubmit"
-          >
-            발급하기
-          </danji-button>
-        </div>
+        <span class="mt-[1.2rem] Body02 text-Black-1">
+          연말정산 시, 현금과 같은 30% 소득공제가 됩니다! ~~~~ 어쩌구 ~~~ ㅓㅈ쩌구 연말정산 시,
+          현금과 같은 30% 소득공제가 됩니다! ~~~~ 어쩌구 ~~~ ㅓㅈ쩌구
+        </span>
+      </div>
+
+      <!-- 하단 버튼 -->
+      <div class="flex justify-center mt-[3.4rem] mb-[4rem]">
+        <danji-button variant="large">발급하기</danji-button>
       </div>
     </template>
-  </Layout>
+  </layout>
 </template>
